Move rules/sbom.bzl to rules_gathering/generate_sbom.bzl
This is part of a continuing cleanup to make license and package
metadata declarations distinct from the rules which create SBOMs
and other reports.

That will make it easier for the declarations to have global
consistency, while individual organizations can define their
own SBOM creators based on local compliance constraints.

Forwarders for moved files are left in place. They will be deleted
by 0.1.0 at the latest.