Move licenses_used from rules/compliance.bzl to sample_reports/licenses_used.bzl
licenses_used is really an example of how the rules can work. As such it is really to be used for examples and as a reference implementation for people to build their own reports from.

This is part of the #85 cleanup.

Bonus: Fix the rule so it does not do a double write.